refactor(types): move types to common module
